fix: improve organization of test opts
